+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/BasicGlobalScopeServices.java
@@ -1,0 +1,111 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.service.scopes;
+
+import org.gradle.api.internal.DocumentationRegistry;
+import org.gradle.api.internal.file.DefaultFileLookup;
+import org.gradle.api.internal.file.FileLookup;
+import org.gradle.api.internal.file.FileResolver;
+import org.gradle.api.tasks.util.PatternSet;
+import org.gradle.api.tasks.util.internal.CachingPatternSpecFactory;
+import org.gradle.api.tasks.util.internal.PatternSets;
+import org.gradle.api.tasks.util.internal.PatternSpecFactory;
+import org.gradle.cache.internal.DefaultFileLockManager;
+import org.gradle.cache.internal.DefaultProcessMetaDataProvider;
+import org.gradle.cache.internal.FileLockManager;
+import org.gradle.cache.internal.locklistener.DefaultFileLockContentionHandler;
+import org.gradle.cache.internal.locklistener.FileLockContentionHandler;
+import org.gradle.internal.Factory;
+import org.gradle.internal.concurrent.DefaultExecutorFactory;
+import org.gradle.internal.concurrent.ExecutorFactory;
+import org.gradle.internal.event.DefaultListenerManager;
+import org.gradle.internal.event.ListenerManager;
+import org.gradle.internal.jvm.inspection.CachingJvmVersionDetector;
+import org.gradle.internal.jvm.inspection.DefaultJvmVersionDetector;
+import org.gradle.internal.jvm.inspection.JvmVersionDetector;
+import org.gradle.internal.nativeintegration.ProcessEnvironment;
+import org.gradle.internal.nativeintegration.filesystem.FileSystem;
+import org.gradle.internal.remote.internal.inet.InetAddressFactory;
+import org.gradle.internal.remote.services.MessagingServices;
+import org.gradle.process.internal.DefaultExecActionFactory;
+import org.gradle.process.internal.ExecHandleFactory;
+
+/**
+ * Defines the basic global services of a given process. This includes the Gradle CLI, daemon and tooling API provider. These services
+ * should be as few as possible to keep the CLI startup fast. Global services that are only needed for the process running the build should go in
+ * {@link GlobalScopeServices}.
+ */
+public class BasicGlobalScopeServices {
+    FileLockManager createFileLockManager(ProcessEnvironment processEnvironment, FileLockContentionHandler fileLockContentionHandler) {
+        return new DefaultFileLockManager(
+            new DefaultProcessMetaDataProvider(
+                processEnvironment),
+            fileLockContentionHandler);
+    }
+
+
+    DefaultFileLockContentionHandler createFileLockContentionHandler(ExecutorFactory executorFactory, InetAddressFactory inetAddressFactory) {
+        return new DefaultFileLockContentionHandler(
+            executorFactory,
+            inetAddressFactory);
+    }
+
+    ExecutorFactory createExecutorFactory() {
+        return new DefaultExecutorFactory();
+    }
+
+    InetAddressFactory createInetAddressFactory(MessagingServices messagingServices) {
+        return messagingServices.get(InetAddressFactory.class);
+    }
+
+    MessagingServices createMessagingServices() {
+        return new MessagingServices();
+    }
+
+    DocumentationRegistry createDocumentationRegistry() {
+        return new DocumentationRegistry();
+    }
+
+    JvmVersionDetector createJvmVersionDetector(ExecHandleFactory execHandleFactory) {
+        return new CachingJvmVersionDetector(new DefaultJvmVersionDetector(execHandleFactory));
+    }
+
+    DefaultExecActionFactory createExecActionFactory(FileResolver fileResolver) {
+        return new DefaultExecActionFactory(fileResolver);
+    }
+
+    FileResolver createFileResolver(FileLookup lookup) {
+        return lookup.getFileResolver();
+    }
+
+    FileLookup createFileLookup(FileSystem fileSystem, Factory<PatternSet> patternSetFactory) {
+        return new DefaultFileLookup(fileSystem, patternSetFactory);
+    }
+
+    PatternSpecFactory createPatternSpecFactory() {
+        return new CachingPatternSpecFactory();
+    }
+
+    protected Factory<PatternSet> createPatternSetFactory(final PatternSpecFactory patternSpecFactory) {
+        return PatternSets.getPatternSetFactory(patternSpecFactory);
+    }
+
+    ListenerManager createListenerManager() {
+        return new DefaultListenerManager();
+    }
+}
+
